Add ability to import `list` w/o `validate`
